working impl with nodes

+++ src/storage/constant.rs
@@ -1,7 +1,7 @@
 // row constants
 pub const USER_NAME_SIZE: usize = 32;
 pub const EMAIL_SIZE: usize = 255;
-pub const ID_SIZE: usize = std::mem::size_of::<u32>();
+pub const ID_SIZE: usize = std::mem::size_of::<usize>();
 pub const ROW_SIZE: usize = USER_NAME_SIZE + EMAIL_SIZE + ID_SIZE;
 // table constants
 pub const PAGE_SIZE: usize = 4096;
